Revert "wip"
This reverts commit b157f57f2aeafe75e3bcbd15ed4684e16201549b.

+++ Elektroauto ESP VSCode/src/mpuFunctions.h
@@ -1,3 +1,3 @@
 void handleMPU();
 void initMPU();
-double rawAccelToPhysicalAccel(int acceleration);
+float rawAccelToPhysicalAccel(int acceleration);
